refactor: 폰트 스타일 사용방식 변경
